[FIX] SortFlex: Adjust customAggregation to include sort order
Change 6922679 introduced the sort order in the affectedControl to ensure the condenser to differentiate between different sort orders.
This introduced an unfortunate side effect: If there is an existing change with a sort direction (e.g. key user adaptation), applying a new one and condensing leads to no condensing. This is due to the original UI state during UI reconstruction having no sorting direction at all.

Therefore, to circumvent this, customAggregation now includes the sorting direction as well, to ensure that condensing can match the correct changes.

SNOW: DINC0637328
Change-Id: I8babac86486e91d1af4940a3a5c73d930cf6da94