[INTERNAL][FIX] sap.ui.fl.diagnostics: Clone flex objects before anonymization
When collecting flexibility data for support diagnostics, the flex
objects cannot be cloned as they must retain their methods for anaylsis
in the flex support extension and are thus not serializable.

This is different in the data extractor for the diagnostics plugin,
where the output is mutated to anonymize sensitive data.
To avoid modifying the original flex objects they must be cloned.
This is possible in the disagnostics plugin because objects are made
serializable during the data transfer process in the plugin API.

SNOW: DINC0693661
SNOW: DINC0693679
Change-Id: I5e01d7d60341553efaee38fc73ab2da7f71f3d9a