[INTERNAL] Remove deprecated code
The test page used deprecated properties that had
no effect anyway because the navButtonText on the
sap.m.Page only effects a specific theme and
requires a value that differs from the default "Back".

Since this usage of deprecated properties is an
issue for the legacy free build, it has been removed.

Change-Id: I921ca6a901cd5d0caf29f3b1097e1146df1d1420