[FIX] sap.uxap.ObjectPageLayout: Fix wrong positioning
Problem: When user navigates from OneColumn layout of FCL to TwoColumns
layout and there is an ObjectPage in the newly visible column, the
scrollable wrapper of ObjectPage is displaced (padding-top is bigger
than it should be).

Solution: Problem happens because header title height is measured during
the resizing animation of the FCL:
- FE call _triggerVisibleSubSectionsEvents
- _requestAdjustLayout is called ->
  _updateScreenHeightSectionBasesAndSpacer is called ->
  _adjustTitlePositioning is called
- padding-top of scrollable wrapper is updated depending on the current
  height of the header title (animation of the FCL is still ongoing,
  title area is bigger as it has wrapping titles, which expand the area,
  when the screen size has small width)

Implemented fix:
When FCL is executing the resize animation, it suspends all resize
listeners of everything inside the resized column.
That's why now in _adjustTitlePositioning function we check if the
resize handler of the header title dom element is suspended and if so -
return and do not update the padding-top of the scrollable wrapper,
based on intermittent title height.
Padding-top is still updated when the resize hanlders are resumed.

SNOW: DINC0686447
Change-Id: I16f8304c11af7c42ff79a72b2b68209130b3a97e