[INTERNAL] sap.ui.fl: replace forceInvalidation with loaderCacheKey in FlexState
The Loader response now generates a loaderCacheKey parameter which is used
to identify changes in the loader data instead of a forceInvalidation boolean flag.
This way a change in the Loader data can be identified in a different thread.
The FlexState also saves this key and compares it when loading data again.

JIRA: TEAM42-475
Change-Id: I7f5803f7c55b82d1a4e4512891d8ff9b285f60f4