[FIX] sap.m.MenuButton: fixed menuPosition functionality
Issue: After the refactoring of the sap.m.Menu, the function of the
menuPosition property of the sap.m.MenuButton was broken because of
different positioning method of the sap.m.ResponsivePopover control
(used now in the Menu) compared to previously used sap.ui.core.Popup.

Solution: Now there is additional Menu popover appearance correction
that restores proper functioning of the menuPosition property.

Snippix for test: #604210

SNOW: DINC0659783
Change-Id: I231c7f6a42e104f6187af2f0ccad165c100a2aae