[FIX] We now trigger a fetch call to load PDF content and determine PDF availability in the PDFViewer
PDFViewer uses iFrame to load the content for the specified source,
However modern browsers implement strict policies for validating external resources loaded within an iframe. PDFViewer cannot determine whether the resource inside the iframe is a valid PDF by itself.

When the source is set, the PDFViewer now automatically validates the resource using a GET request to ensure it exists and is accessible.

This validation are as below
Prevents loading invalid or non-existent PDF files
If validation fails, error content is displayed instead of attempting PDF load

SNOW: DINC0680978
Change-Id: Iad075dbd8fb59d3f5000308c7e93ed847240ec43