[INTERNAL] sap.ui.rta: Set max. width for cba MenuButton in RTA
In RTA, when context-based adaptation is enabled, a menu button is
rendered in the RTA which displays the title of the currently adapted
context-based adaptation. We allow context-based adaptations to have a
title length of 100. Because the menu button has no max-width property,
the whole window width is used to display the context-based adaptation
title. This will be prevent with this change, by setting a max-width.

Jira: UITECHGRAVITY-5642
Change-Id: Id29ab601d485127092d300b8d85ba5dfbc682af7