[FIX] sap/ui/model/ClientTreeBinding#cloneData: use a clone for this.oTreeData
In order to correctly determine if a change is went needs to be fired
the stored tree data and the new tree data need to be compared.
If they both point to the same object reference, no changes are detected.
Therefore we need to create a clone of the tree data in cloneData

PS1: Integration test, code, TDD
PS2: override cloneData in XMLTreeBinding, add test
PS3-4: update commit message
PS5: XMLTreeBinding#cloneData returns symbol and thus always fires a
   change event when #checkUpdate is called.

Change-Id: I72816121789662a0ad4eb2dc5481faff77f71d40
SNOW: DINC0701989