[FIX] sap.uxap.ObjectPageSection: Fix sticky headers when title on left
Problem: When ObjectPageLayout has subSectionLayout="TitleOnLeft" headers of both Section and SubSection are displayed (by design) even when there is only one SubSection in a given Section. In this case, on bigger screen sizes (Desktop and Large Desktop) when titles of SubSections are indeed displayed on the left , both headers of Section and SubSection are now sticky, which is not a correct behavior. Only one header should be sticky (most specific one - in this case this is the header of the SubSection). Now upon scrolling down - the Section header overlaps some part of the SubSection content and a part of it is also being visible below the header of the SubSection at some point.

Solution: Make the header of the Section in this case (only one
SubSection in a given Section with both titles visible and title on left) not sticky.

SNOW: DINC0706154
Change-Id: Ie29a369c46a3e4f1a16c3f01462b9e39da9b2f76