[FIX] We now ensure that the UploadSetWithTable plugin correctly handles file extraction while maintaining the original MIME type when you drag and drop files into the plugin
Issue: The Datatransfer item's webkitEntry was not fetching the mime type of file dragged and dropped into the plugin as it relies on the browser mime sniffing ability, which the browsers like chrome does not sniff the mimetype.

resolution: To extract the file without losing the mime type we now also pass the dataTransferItem entry along with webkitEntry so we extract from the datatransferItem if it is a file which retains the file's type.

SNOW: DINC0677220
Change-Id: I002cc27e8b55b1a0e2cdde2b189bbd4f5c192be7