[FIX] sap.uxap.ObjectPageSubSection: fix height of header
Problem: When enabling wrapping titles on ObjectPageSubSection, height
auto was introduced for header toolbar, but in this case, if no action
buttons are provided, header is too small (height of the title itself).

Solution: Still keep height: auto of the header toolbar, but set
min-height 2.75rem (desired height as per VD spec).

SNOW: N/A
Change-Id: I7248a3a72e3313cefc45ee350224a65b7ee30019