[FIX] sap.f.DynamicPage: update title positioning
Problem: Sticky header of table, placed in DynamicPage, gets overlapped
by sticky header provider (IconTabBar), when user scrolls down -> header
of table looks like it does not stick.

Solution: This happens when DynamicPage does not have header. When user
scrolls down and the header should snap and the sticky content should go
to the title area (in sticky place holder), we do not update the
padding-top of the scrollable container (to add the height of the new
sticky content). As a result, sticky header of the table is scrolled
below the sticky header provider of the DynamicPage.

Now the title positioning is updated, no matter DynamicPage does not
have header.

SNOW: DINC0691965
Change-Id: I153f98b894f0b6043cf9161637d4d97d0eceb6d2