[FIX] sap.m.TimePicker: fixed Dialog title on mobile devices
Issue: there was incorrectly working functionality for setting or
clearing the title of the picker Dialog on mobile devices, and also
there was no default label for this dialog in case there is no
associated label or title set.

Solution: now a default label is added, and the proper title is being
displayed depending on associated label existence or explicitly set
title.

SNOW: DINC0712021
Change-Id: I156dba1c6af127a93ef041f3e618094cbf5ddba3