[INTERNAL][FIX] sap.ui.rta: Fix iframe url binding leak
When an iframe url contains a binding string we need to resolve the
actual property value in the iframe dialog to show a correct preview.
For this, a hidden helper control is used, where a property is bound
using the same context as the iframe control to later fetch the
resolved value.
Bindings against the ODataModel were causing issues as they are async
and two-way by default.

When subsequent url updates after an initial binding introduce a new
binding, the old binding gets cleaned up automatically by the core.
However if the update is a plain string, the value will actually flow
back into the model property via the old two-way binding as it was never
cleaned up.

Since the binding is resolved async it cannot be cleaned up immediately.
Also, as bindings in the iframe context can be complex (multiple parts,
expression bindings etc.) we cannot override the binding mode in an
easy and risk-free way.

As a solution we now clean up old bindings explicitly whenever receiving
a new value that is not a binding.
This can be checked via the BindingParser, as it will only return
extracted binding info, thus not returning anything for plain values.

SNOW: DINC0694603
Change-Id: Ie67493eaba720513e4c688f877e8fdd83be89670