[INTERNAL] Table: Improve selection update on scroll
The responsibility to update the header selector state is moved from the table to the selection plugins.
The table has to request the current state on every update. This can lead to many computations of the current state, for example when scrolling. Plugins can update the header selector state only when needed.

- Added HeaderSelector control to manage header selection visualization in tables. The table manages the lifecycle of the HeaderSelector, renders it in the correct place, and provides access to selection plugins so they can update its properties when required.

- Modified selection plugins to integrate with HeaderSelector, ensuring correct visibility and state management.

- In the ODataV4MultiSelectionPlugin, updates are only done when the plugin detects that the state might have changed. Other plugins might still do many, perhaps unnecessary, updates. These were so far not considered performance critical.

Change-Id: Ib471e6df37c234e65fa1a7ca12c5c3ad7d40dfb9
Jira: CPOUIFTEAMB-2583