[FIX] sap.m.MenuButton: Fix MenuButton menuPosition to handle viewport collision detection
Issue: The previous fix (Change ID 7034988) restored the menuPosition
functionality but didn't handle cases when the whole menu or part of
it appeared outside of the viewport.

Solution: Now there is collision detection and automatic position
adjustment for the Menu to prevent it from appearing offscreen.

SNOW: DINC0709854
SNOW: DINC0722151
SNOW: DINC0727560
Change-Id: I6831223ef774abd36e84bf33c530a9854c0a9ede