[INTERNAL] mdc.FieldBase: do not suppress invalidation for visible property
Issue: FieldBase#setProperty suppresses the invalidation for properties that are rendered on the content control. This includes the "visible" property. This lead to the issue that bound values for the MultiValueField's "visible" property did not update this property although the model was

Fix: remove the visible property from being suppressed.

SNOW: DINC0735786
Change-Id: Ic51b5263062db24f28bf30587238859cbca83e73