[INTERNAL][FIX]sap.ui.fl: Ensure FlexState is initialized if data was previously cached
A saved draft can contain all existing changes for an app. When such a
draft is discarded, the FlexState must be initialized again to clean up
the state and remove those changes. Previously, because no changes were
retrieved from the Loader, we assumed a re-initialization was not needed.

This commit adds a flag to indicate whether relevant data was previously
cached. The ComponentLifecycleHooks then uses this flag to determine
whether the FlexState needs to be initialized again.

SNOW: DINC0743360
Change-Id: I6cc8c832b3bc6859713cecfcf206027c1dada366