[INTERNAL][FEATURE] JSDoc: automatically add @lends tag to extend calls
When an extend call creates a class and when the name of the class is
known and when the extend call's second argument is an object literal,
then ensure that it has a @lends tag associating the object literal
with the class prototype.

Also logs a future error when the name of an existing @lends tags
doesn't match the name of the described class.

Change-Id: I51b19b213f3bf8c6641e51a19a11771a2a5fd784